requirements.txt
openenv-core
fastapi
uvicorn[standard]
pydantic
openai
requests
huggingface_hub